refactor(utils): extract manual edit conflict logic into separate utility function
The logic for handling manual edit conflicts in the `CodeAndPreview` component was moved to a new utility function `informManualEditConflicts` to improve code maintainability and reusability. This change also involved removing the deprecated `handleCustomEditsRemoval` function and exporting the `ToasterToast` type for use in the new utility.
